+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### 2.0.1
+- [PR#102](https://github.com/twreporter/go-api/pull/102)
+- Use userID, email and standard claims to generate JWT.
+- Code refactors. Fix typo, add error check.
+- Still redirect to destination even if oauth fails.
+- New endpoint "/v1/token/:userID", which is used to renew JWT for clients.
+- Set Cache-Control: no-store for those endpoints related to users
+
 ### 2.0.0
 **Major Change**
 - Drop password and signup process, only send activation email every time user want to sign in.